Pavel Roskin <proski@gnu.org>:
The current CVS version of ignores ESSID settings.  When ESSID is set
using iwconfig, the device reports that it's disconnected, and then
associates to a random AP in the range, even it its ESSID is different.

The simplest test is to set ESSID to a non-existing string (e.g.
"non-existing") and see it associate.  It happens whenever the interface
is up.  ESSID can be set when the interface is up or down.

The fix it to set ESSID length properly