More cleanups
Change the way we are finding the kernel tree and the kernel version.
Rename KERNEL_SRC to KERNEL_PATH because we actually need the binary
tree, which is not always the source tree.

Prepare for Linux 2.6.18 - try reading version from utsrelease.h first.

Fix package version calculation.  We don't need it when called by the
kernel.