Reformat to be more similar to other firmware files
Although a comment in the file says that it was automatically generated
"for AT76C502A", it was likely generated by an older version that would
put commas in the beginning of strings.

No references to for "AT76C502A" can be found in any script or program
currently in CVS, so the file cannot be regenerated by the tool that
generated it.