Join all firmware drivers into one
The new driver is called at76fw.c (the name is provisional, as it's
likely to be joined with at76c503.c soon).  Firmware linking into the
drivers is now disabled.  "make dist" should build only one tarball, the
one without the headers.  Nothing needs to be appended to the version.